extract sil body lowering
