+++ src/compiler/codegen/sil/mod.rs
@@ -1,1 +1,4 @@
-pub(super) mod token_refs;
+mod body;
+mod token_refs;
+
+pub(super) use body::{lower_entry_body, lower_entry_expr};
